Move afterRunHander to IPC testRunner/
